Move caching e2e tests to separate file
